refactor: reduce duplication and resolve merge conflicts
