+++ src/main/java/uk/gov/hmcts/reform/civil/handler/callback/camunda/notification/DefendantResponseCaseHandedOfflineApplicantNotificationHandler.java
@@ -11,18 +11,14 @@ import uk.gov.hmcts.reform.civil.callback.CaseEvent;
 import uk.gov.hmcts.reform.civil.config.properties.notification.NotificationsProperties;
 import uk.gov.hmcts.reform.civil.model.CaseData;
 import uk.gov.hmcts.reform.civil.service.NotificationService;
+import uk.gov.hmcts.reform.civil.utils.NotificationUtils;
 
 import java.util.List;
 import java.util.Map;
 
 import static uk.gov.hmcts.reform.civil.callback.CallbackType.ABOUT_TO_SUBMIT;
 import static uk.gov.hmcts.reform.civil.callback.CaseEvent.NOTIFY_APPLICANT_SOLICITOR1_FOR_CASE_HANDED_OFFLINE;
-import static uk.gov.hmcts.reform.civil.enums.MultiPartyScenario.ONE_V_ONE;
-import static uk.gov.hmcts.reform.civil.enums.MultiPartyScenario.TWO_V_ONE;
-import static uk.gov.hmcts.reform.civil.enums.MultiPartyScenario.getMultiPartyScenario;
 import static uk.gov.hmcts.reform.civil.utils.NotificationUtils.is1v1Or2v1Case;
-import static uk.gov.hmcts.reform.civil.utils.PartyUtils.buildPartiesReferences;
-import static uk.gov.hmcts.reform.civil.utils.PartyUtils.getPartyNameBasedOnType;
 
 @Service
 @RequiredArgsConstructor
@@ -79,31 +75,6 @@ public class DefendantResponseCaseHandedOfflineApplicantNotificationHandler exte
 
     @Override
     public Map<String, String> addProperties(CaseData caseData) {
-        if (getMultiPartyScenario(caseData).equals(ONE_V_ONE)) {
-            return Map.of(
-                CLAIM_REFERENCE_NUMBER, caseData.getLegacyCaseReference(),
-                REASON, caseData.getRespondent1ClaimResponseType().getDisplayedValue(),
-                PARTY_REFERENCES, buildPartiesReferences(caseData)
-            );
-        } else if (getMultiPartyScenario(caseData).equals(TWO_V_ONE)) {
-            return Map.of(
-                CLAIM_REFERENCE_NUMBER, caseData.getLegacyCaseReference(),
-                REASON, caseData.getRespondent1ClaimResponseType().getDisplayedValue()
-                    .concat(" against " + caseData.getApplicant1().getPartyName())
-                    .concat(" and " + caseData.getRespondent1ClaimResponseTypeToApplicant2())
-                    .concat(" against " + caseData.getApplicant2().getPartyName()),
-                PARTY_REFERENCES, buildPartiesReferences(caseData)
-            );
-        } else {
-            //1v2 template is used and expects different data
-            return Map.of(
-                CLAIM_REFERENCE_NUMBER, caseData.getLegacyCaseReference(),
-                RESPONDENT_ONE_NAME, getPartyNameBasedOnType(caseData.getRespondent1()),
-                RESPONDENT_TWO_NAME, getPartyNameBasedOnType(caseData.getRespondent2()),
-                RESPONDENT_ONE_RESPONSE, caseData.getRespondent1ClaimResponseType().getDisplayedValue(),
-                RESPONDENT_TWO_RESPONSE, caseData.getRespondent2ClaimResponseType().getDisplayedValue(),
-                PARTY_REFERENCES, buildPartiesReferences(caseData)
-            );
-        }
+        return NotificationUtils.caseOfflineNotificationAddProperties(caseData);
     }
 }
